fix: revalidate probe paths and separate platform fixture suites
